feat: added font custom variations
