+++ package.json
@@ -76,6 +76,7 @@
     "ts-jest": "28.0.3",
     "ts-node": "10.8.0",
     "typescript": "4.7.2",
+    "url-loader": "^4.1.1",
     "webpack": "5.72.1"
   },
   "dependencies": {
