+++ src/components/themeProvider/index.tsx
@@ -1,8 +1,9 @@
 import React, { FC, PropsWithChildren } from 'react';
 import { ChakraProvider, CSSReset } from '@chakra-ui/react';
 
-import { Fonts } from './fonts';
 import { autoScout24Theme, motoScout24Theme } from '../../theme';
+// eslint-disable-next-line import/no-internal-modules
+import { Fonts } from '../../styles/fonts';
 
 export type Props = {
   // Theme to use
