+++ tsconfig.json
@@ -17,5 +17,5 @@
     "esModuleInterop": true
   },
   "exclude": ["dist", "node_modules"],
-  "include": ["src"]
+  "include": ["src", "typings/**/*.d.ts"],
 }
