+++ typings/fonts.d.ts
@@ -1,0 +1,3 @@
+declare module '*.otf';
+declare module '*.woff';
+declare module '*.woff2';
